reuse code for testnet+mainnet
