+++ typescript/infra/config/environments/testnet4/igp.ts
@@ -1,23 +1,14 @@
-import { BigNumber, utils as ethersUtils } from 'ethers';
-
-import {
-  ChainMap,
-  ChainName,
-  HookType,
-  IgpConfig,
-  TOKEN_EXCHANGE_RATE_DECIMALS,
-  defaultMultisigConfigs,
-  multisigIsmVerificationCost,
-} from '@hyperlane-xyz/sdk';
+import { ChainMap, HookType, IgpConfig } from '@hyperlane-xyz/sdk';
 import { Address, exclude, objMap } from '@hyperlane-xyz/utils';
 
 import {
   AllStorageGasOracleConfigs,
   getAllStorageGasOracleConfigs,
   getTokenExchangeRateFromValues,
+  remoteOverhead,
 } from '../../../src/config/gas-oracle.js';
-import { isEthereumProtocolChain } from '../../../src/utils/utils.js';
 
+import { ethereumChainNames } from './chains.js';
 import gasPrices from './gasPrices.json';
 import { owners } from './owners.js';
 import { supportedChainNames } from './supportedChainNames.js';
@@ -25,36 +16,14 @@ import rawTokenPrices from './tokenPrices.json';
 
 const tokenPrices: ChainMap<string> = rawTokenPrices;
 
-const FOREIGN_DEFAULT_OVERHEAD = 600_000; // cosmwasm warp route somewhat arbitrarily chosen
-const remoteOverhead = (remote: ChainName) =>
-  supportedChainNames.filter(isEthereumProtocolChain).includes(remote as any)
-    ? multisigIsmVerificationCost(
-        defaultMultisigConfigs[remote].threshold,
-        defaultMultisigConfigs[remote].validators.length,
-      )
-    : FOREIGN_DEFAULT_OVERHEAD; // non-ethereum overhead
-
-// Gets the exchange rate of the remote quoted in local tokens
-function getTokenExchangeRate(local: ChainName, remote: ChainName): BigNumber {
-  const localValue = ethersUtils.parseUnits(
-    tokenPrices[local],
-    TOKEN_EXCHANGE_RATE_DECIMALS,
-  );
-  const remoteValue = ethersUtils.parseUnits(
-    tokenPrices[remote],
-    TOKEN_EXCHANGE_RATE_DECIMALS,
-  );
-
-  return getTokenExchangeRateFromValues(local, localValue, remote, remoteValue);
-}
-
 export const storageGasOracleConfig: AllStorageGasOracleConfigs =
   getAllStorageGasOracleConfigs(
     supportedChainNames,
     gasPrices,
-    getTokenExchangeRate,
+    (local, remote) =>
+      getTokenExchangeRateFromValues(local, remote, tokenPrices),
     (local) => parseFloat(tokenPrices[local]),
-    (local) => remoteOverhead(local),
+    (local) => remoteOverhead(local, ethereumChainNames),
   );
 
 export const igp: ChainMap<IgpConfig> = objMap(
@@ -69,7 +38,7 @@ export const igp: ChainMap<IgpConfig> = objMap(
       overhead: Object.fromEntries(
         exclude(chain, supportedChainNames).map((remote) => [
           remote,
-          remoteOverhead(remote),
+          remoteOverhead(remote, ethereumChainNames),
         ]),
       ),
     };
